test(browser): share the notes page helper and the approval fixture
Four Manager notes smokes each launched Chromium, pushed the browser onto their
own `browsers` array and registered the same after() closer — three under an
identical `pageForManager`, the fourth inline. tests/browser/manager-notes-page.ts
now owns the launch, the registry and the cleanup, so the next file that needs a
page cannot forget to close one.

manager-layout-smoke keeps its own `pageForManager`. Same name, different
contract: it connects to an already-running browser over CDP, takes a
TestContext and skips when none is there. Merging them would mean one of the two
callers getting a browser it did not ask for.

The three dispatch-approval ingress tests carried a byte-identical `pending()`
over dispatchApprovalStore. They assert one approval policy across three
transports, so they now compare against one fixture in
tests/unit/helpers/dispatch-approval.ts. The other `pending()` functions under
tests/unit share only the name and are left alone.
